changes for VerifyDeletionOfSubscribedProduct_Test

+++ src/rhsm/cli/tests/BugzillaTests.java
@@ -2161,20 +2161,22 @@ public class BugzillaTests extends SubscriptionManagerCLITestScript {
 				(List<String>) null, null, null, null, true, null, null, null, null);
 		clienttasks.subscribe_(true, null, null, (String) null, null, null,
 				null, null, null, null, null);
-		for (InstalledProduct installed : clienttasks
-				.getCurrentlyInstalledProducts()) {
+		if(clienttasks.getCurrentlyConsumedProductSubscriptions().isEmpty()){
+			throw new SkipException("no installed products are installed");
+		}
+		else{
+		for (InstalledProduct installed : clienttasks.getCurrentlyInstalledProducts()) {
+			
 			if (installed.status.equals("Subscribed")) {
-				for (SubscriptionPool AvailSub : clienttasks
-						.getCurrentlyAvailableSubscriptionPools()) {
-					if (installed.productName
-							.contains(AvailSub.subscriptionName)) {
+				for (SubscriptionPool AvailSub : clienttasks.getCurrentlyAvailableSubscriptionPools()) {
+					if (installed.productName.contains(AvailSub.subscriptionName)) {
 						String jsonConsumer = CandlepinTasks.deleteResourceUsingRESTfulAPI(sm_serverAdminUsername,sm_serverAdminPassword, sm_serverUrl,"/products/" + AvailSub.productId);
 						String expect = "{\"displayMessage\""+ ":"+ "\"Product with UUID '"+ AvailSub.productId+ "'"+ " cannot be deleted while subscriptions exist.\"}";
 						Assert.assertEquals(expect, jsonConsumer);
 					}
 				}
 			}
-		}
+		}}
 
 	}
